+++ src/app/widgets/svg-autopilot-v2/svg-autopilot-v2.component.scss
@@ -1,0 +1,52 @@
+:host {
+  display: block;
+  width: 100%;
+  font-family: Roboto;
+  overflow: hidden;
+  aspect-ratio: 2 / 1; // width:height = 2:1, so height is half the width
+  max-width: 100%;
+  max-height: 100%;
+}
+
+svg {
+  width: 100%;
+  height: 100%;
+  display: block;
+  overflow: hidden;
+}
+
+.starboard  {
+  stroke: var(--kip-starboard-color);
+}
+
+.port {
+  stroke: var(--kip-port-color);
+}
+
+.contrast-text {
+  fill: var(--kip-contrast-color);
+  stroke: var(--kip-contrast-color);
+}
+
+.contrast-text-dim {
+  fill: var(--kip-contrast-dim-color);
+  stroke: var(--kip-contrast-dim-color);
+}
+
+.appWA-text {
+  fill: var(--kip-orange-color);
+  stroke: var(--kip-orange-color);
+}
+
+.appWA-text-dim {
+  fill: var(--kip-orange-dim-color);
+  stroke: var(--kip-orange-dim-color);
+}
+
+.card-color {
+  fill: var(--kip-widget-card-background-color);
+}
+
+.fixed-dial-background {
+  fill: var(--kip-grey-color);
+}
