--- src/app/widgets/widget-autopilot copy/widget-autopilot.component.scss
@@ -1,163 +1,0 @@
-:root {
-  display: block;
-  height: 100%;
-  width: 100%;
-}
-
-.apStencil.confirmCommand {
-  z-index: 5;
-  visibility: hidden;
-}
-
-.apStencil.error {
-  z-index: 6;
-  visibility: hidden;
-}
-
-.apCompass {
-  position: relative;
-  width: 360px;
-  height: 202px;
-  margin: auto;
-  border-radius: 5px;
-}
-
-div.apStatusBar {
-  width: 360px;
-  height: 18.5px;
-  border-bottom-left-radius: 5px;
-  border-bottom-right-radius: 5px;
-  background-color: red;
-}
-
-.apStatusBar, .apStatusBar DIV {
-  display: inline-block;
-  position: relative;
-}
-
-div.autopilotWidget {
-  display:block;
-  position: relative;
-  border:none;
-  margin: 0px;
-  padding:  5px 0px;
-  width: 100%;
-  height: calc(100% - 10px);
-}
-
-.headerAutopilot {
-  display: block;
-  position: relative;
-  width: 346px;
-  margin: auto;
-  margin-bottom: 5px;
-}
-
-.headerAutopilot .apTitle {
-  display: inline-block;
-  font-weight: bold;
-  width: calc(100% - 64px);
-  margin: 0px;
-}
-
-.headerAutopilot .apPower {
-  display: inline-block;
-  width: 64px;
-  margin: 0px;
-}
-
-.auto-pilot-btn-width {
-  width:85px;
-}
-
-.button-color-red {
-  --mdc-filled-button-container-color: var(--kip-port-color);
-  --mdc-filled-button-label-text-color: var(--kip-contrast-color);
-}
-
-.activityIcon {
-  opacity: 40%;
-  color: red;
-  width:25px;
-}
-
-.autopilotGridBtnContainer {
-  display: grid;
-  width: fit-content;
-  margin: auto;
-  margin-top: 5px;
-  height: 160px;
-  grid-template-columns: [col-start] 25% [col1-end] 25% [col2-end] 25% [col3-end] 25% [col-End];
-  grid-template-rows: [row-start] 25% [row1-end] 25% [row2-end] 25% [row3-end] 25% [row-end];
-  grid-template-areas:
-    ". mute message ."
-    "port minus1  plus1 starboard"
-    "windmode minus10 plus10 trackmode"
-    "standby . . auto1";
-  row-gap: 6px;
-  column-gap: 4px;
-  justify-items: center;
-  align-items: center;
-  justify-content: center;
-  align-content: start;
-  --mdc-filled-button-container-shape: 5px;
-}
-
-.itemStbTack {
-  grid-area: starboard;
-}
-
-.itemMute {
-  grid-area: mute;
-}
-
-.itemMessage {
-  grid-area: message;
-}
-
-.itemPrtTack {
-  grid-area: port;
-}
-
-.itemPlus1 {
-  grid-area: plus1;
-}
-
-.itemMinus1 {
-  grid-area: minus1;
-}
-
-.itemStandby {
-  grid-area: standby;
-}
-
-.itemPlus10 {
-  grid-area: plus10;
-}
-
-.itemMinus10 {
-  grid-area: minus10;
-}
-
-.itemAuto {
-  grid-area: auto1;
-}
-
-.itemWindmode {
-  grid-area: windmode;
-}
-
-.itemTrackmode {
-  grid-area: trackmode;
-}
-
-.itemNxtWp {
-  grid-area: nxtWp;
-}
-
-.counterText {
-  font-weight: bold;
-  font-size: 7em;
-  margin-top: 0px;
-  margin-bottom: 0px;
-}
