+++ src/app/widgets/widget-autopilot/_widget-autopilot.component-theme.scss
