+++ src/app/widgets/widget-autopilot/widget-autopilot.component.scss
@@ -1,224 +1,163 @@
-@use '@angular/material' as mat;
-
-:host {
+:root {
   display: block;
   height: 100%;
   width: 100%;
-  min-height: 0;
-  min-width: 0;
 }
 
-.widget-host {
-  display: block;
-  height: 100%;
-  width: 100%;
-  min-height: 0;
-  min-width: 0;
+.apStencil.confirmCommand {
+  z-index: 5;
+  visibility: hidden;
 }
 
-.autopilot-grid-container,
-.autopilot-grid-row  {
-  min-height: 0;
-  min-width: 0;
+.apStencil.error {
+  z-index: 6;
+  visibility: hidden;
 }
 
-.autopilot-grid-container {
-  display: grid;
-  gap: 0rem 1rem;
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-  padding: 0.25rem 0.75rem 0.25rem 0.75rem;
+.apCompass {
+  position: relative;
+  width: 360px;
+  height: 202px;
+  margin: auto;
+  border-radius: 5px;
 }
 
-.autopilot-grid-container.rows-standby { grid-template-rows: 14fr auto 20fr; }
-.autopilot-grid-container.rows-wind { grid-template-rows: 14fr auto 20fr 14fr 14fr 14fr; }
-.autopilot-grid-container.rows-auto { grid-template-rows: 14fr auto 20fr 14fr 14fr; }
-.autopilot-grid-container.rows-route { grid-template-rows: 1fr auto 1fr 1fr 1fr 1fr; }
-
-
-.autopilot-grid-container
-.autopilot-grid-row {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 1rem;
-  padding: 0.5rem 0rem;
+div.apStatusBar {
+  width: 360px;
+  height: 18.5px;
+  border-bottom-left-radius: 5px;
+  border-bottom-right-radius: 5px;
+  background-color: red;
 }
 
-.autopilot-grid-row--top {
-  justify-content: space-between;
+.apStatusBar, .apStatusBar DIV {
+  display: inline-block;
+  position: relative;
 }
 
-.autopilot-btn {
-  flex: 1 1 0;
-  width: 100%;
-  border-radius: 12px;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-blue-dim-color),
-    filled-container-height: 100%
-  ));
-}
-
-.autopilot-btn--modes {
-  @include mat.button-overrides((
-    filled-horizontal-padding: 0rem
-  ));
-}
-
-.autopilot-btn--engage {
-  @include mat.button-overrides((
-    filled-horizontal-padding: 0rem,
-    filled-container-color: var(--kip-zone-alarm-color),
-  ));
-}
-
-.autopilot-grid-row--rudder,
-.autopilot-grid-row--autopilot {
-  width: 100%;
-  justify-content: center;
-  padding-bottom: 0rem;
-}
-
-.autopilot-grid-row--rudder {
-  padding: 0rem;
-}
-
-.autopilot-grid-row--autopilot {
-  height: 100%;
-  padding-top: 0rem;
-}
-
-.autopilot-grid-row--actions {
-  justify-content: space-evenly;
-}
-
-@media (max-width: 600px) {
-  .autopilot-grid-container {
-    padding: 0.5rem;
-    gap: 0.25rem;
-  }
-
-  .autopilot-btn {
-    font-size: 0.95rem;
-    padding: 0.4em 0.5em;
-  }
-
-  .autopilot-btn--modes,
-  .autopilot-btn--engage {
-    max-width: 100%;
-  }
-
-  .autopilot-grid-row--top {
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-}
-
-.svg-icon {
-  width: 100%;
-  height: 100%;
+div.autopilotWidget {
+  display:block;
+  position: relative;
+  border:none;
   margin: 0px;
-  padding: 0.5rem 0px;
+  padding:  5px 0px;
+  width: 100%;
+  height: calc(100% - 10px);
 }
 
-.svg-element {
-  width: 100%;
-  height: 100%;
-  text {
-    font-family: Roboto;
-    font-size: 25px;
-    font-weight: bold;
-    text-align: center;
-    text-anchor: middle;
-    font-style: normal;
-    font-variant: normal;
-    font-stretch: normal;
-    font-variant-ligatures: normal;
-    font-variant-caps: normal;
-    font-variant-numeric: normal;
-    font-variant-east-asian: normal;
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: none;
-  }
-  path {
-    fill: none;
-    fill-opacity: 0;
-    stroke: var(--kip-contrast-color);
-    stroke-opacity: 1;
-    stroke-width :7.59058;
-    stroke-linecap: round;
-    stroke-miterlimit: 0;
-    stroke-dasharray: none;
-  }
-  path.tack {
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: var(--kip-contrast-color);
-    stroke-opacity: 1;
-  }
+.headerAutopilot {
+  display: block;
+  position: relative;
+  width: 346px;
+  margin: auto;
+  margin-bottom: 5px;
 }
 
-.svg-modes {
-  width: 100%;
-  height: 100%;
-  text {
-    font-family: Roboto;
-    font-size: 11px;
-    font-weight: bold;
-    text-align: center;
-    text-anchor: middle;
-    font-style: normal;
-    font-variant: normal;
-    font-stretch: normal;
-    font-variant-ligatures: normal;
-    font-variant-caps: normal;
-    font-variant-numeric: normal;
-    font-variant-east-asian: normal;
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: none;
-  }
+.headerAutopilot .apTitle {
+  display: inline-block;
+  font-weight: bold;
+  width: calc(100% - 64px);
+  margin: 0px;
 }
 
-.svg-route {
-  width: 100%;
-  height: 100%;
-  text {
-    font-family: Roboto;
-    font-size: 25px;
-    font-weight: bold;
-    text-align: center;
-    text-anchor: middle;
-    font-style: normal;
-    font-variant: normal;
-    font-stretch: normal;
-    font-variant-ligatures: normal;
-    font-variant-caps: normal;
-    font-variant-numeric: normal;
-    font-variant-east-asian: normal;
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: none;
-  }
-  path {
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: var(--kip-contrast-color);
-    stroke-opacity: 1;
-    stroke-width :7.59058;
-    stroke-linecap: round;
-    stroke-miterlimit: 0;
-    stroke-dasharray: none;
-  }
-  path.tack {
-    fill: var(--kip-contrast-color);
-    fill-opacity: 1;
-    stroke: var(--kip-contrast-color);
-    stroke-opacity: 1;
-  }
+.headerAutopilot .apPower {
+  display: inline-block;
+  width: 64px;
+  margin: 0px;
+}
+
+.auto-pilot-btn-width {
+  width:85px;
+}
+
+.button-color-red {
+  --mdc-filled-button-container-color: var(--kip-port-color);
+  --mdc-filled-button-label-text-color: var(--kip-contrast-color);
+}
+
+.activityIcon {
+  opacity: 40%;
+  color: red;
+  width:25px;
+}
+
+.autopilotGridBtnContainer {
+  display: grid;
+  width: fit-content;
+  margin: auto;
+  margin-top: 5px;
+  height: 160px;
+  grid-template-columns: [col-start] 25% [col1-end] 25% [col2-end] 25% [col3-end] 25% [col-End];
+  grid-template-rows: [row-start] 25% [row1-end] 25% [row2-end] 25% [row3-end] 25% [row-end];
+  grid-template-areas:
+    ". mute message ."
+    "port minus1  plus1 starboard"
+    "windmode minus10 plus10 trackmode"
+    "standby . . auto1";
+  row-gap: 6px;
+  column-gap: 4px;
+  justify-items: center;
+  align-items: center;
+  justify-content: center;
+  align-content: start;
+  --mdc-filled-button-container-shape: 5px;
+}
+
+.itemStbTack {
+  grid-area: starboard;
+}
+
+.itemMute {
+  grid-area: mute;
+}
+
+.itemMessage {
+  grid-area: message;
+}
+
+.itemPrtTack {
+  grid-area: port;
+}
+
+.itemPlus1 {
+  grid-area: plus1;
+}
+
+.itemMinus1 {
+  grid-area: minus1;
+}
+
+.itemStandby {
+  grid-area: standby;
+}
+
+.itemPlus10 {
+  grid-area: plus10;
+}
+
+.itemMinus10 {
+  grid-area: minus10;
+}
+
+.itemAuto {
+  grid-area: auto1;
+}
+
+.itemWindmode {
+  grid-area: windmode;
+}
+
+.itemTrackmode {
+  grid-area: trackmode;
+}
+
+.itemNxtWp {
+  grid-area: nxtWp;
+}
+
+.counterText {
+  font-weight: bold;
+  font-size: 7em;
+  margin-top: 0px;
+  margin-bottom: 0px;
 }
