+++ src/app/widgets/widget-numeric/widget-numeric.component.scss
@@ -1,3 +1,9 @@
+:host {
+  display: block;
+  width: 100%;
+  height: 100%;
+}
+
 .textGenericWrapper {
   position: absolute;
   display: block;
